rc-fusion/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
gitdb2==2.0.2             # via gitpython
gitpython==2.1.5
jinja2==2.9.6
markupsafe==1.0           # via jinja2
pyyaml==3.12
smmap2==2.0.3             # via gitdb2
